Support ng-template tooltips (Needs more work)
They're showing, and tracking but not in the right position I think.
Needs some more work, and thinking about what we actually need. If you
don't need an ng-template the toolTipFormatter should be enough as it
can render html. So, maybe we should just revert this? Worth playing
around with maybe?