Delay scopes check - cleanup. Fixes #4213 in some cases
Scopes will now only be loaded after the first request to Kavita. And
OIDC will not fail if scopes can't be loaded, this is only an issue for
one provider. But if Kavita can't reach it, you likely can't either so I
suppose it's fine.

This will fix issues where users didn't have their idp accessibly right
at the start of Kavita